Abstract common code into base class
