+++ test/hotspot/jtreg/runtime/cds/appcds/jcmd/JCmdTestDynamicDump.java
@@ -29,7 +29,7 @@
  * @requires vm.cds
  * @library /test/lib /test/hotspot/jtreg/runtime/cds/appcds
  * @modules jdk.jcmd/sun.tools.common:+open
- * @compile ../test-classes/Hello.java
+ * @compile ../test-classes/Hello.java JCmdTestDumpBase.java
  * @build sun.hotspot.WhiteBox
  * @build JCmdTestLingeredApp JCmdTestDynamicDump
  * @run driver jdk.test.lib.helpers.ClassFileInstaller sun.hotspot.WhiteBox
@@ -47,145 +47,17 @@ import jdk.test.lib.dcmd.PidJcmdExecutor;
 import jdk.test.lib.process.OutputAnalyzer;
 import jdk.test.lib.process.ProcessTools;
 import jdk.test.lib.JDKToolFinder;
-import jtreg.SkippedException;
-import sun.hotspot.WhiteBox;
 
 import java.io.InputStreamReader;
 import java.io.BufferedReader;
 
-public class JCmdTestDynamicDump {
-    static final String TEST_CLASSES[]      = {"JCmdTestLingeredApp",
-                                               "jdk/test/lib/apps/LingeredApp",
-                                               "jdk/test/lib/apps/LingeredApp$1"};
-    static final String BOOT_CLASSES[]      = {"Hello"};
-
+public class JCmdTestDynamicDump extends JCmdTestDumpBase {
     static final String DYNAMIC_DUMP_FILE   = "mydynamic";
-
-
     static final String[] DYNAMIC_MESSAGES  = {"JCmdTestLingeredApp source: shared objects file (top)",
                                                "LingeredApp source: shared objects file (top)",
                                                "Hello source: shared objects file (top)"};
-
-    static String testJar = null;
-    static String bootJar = null;
-    static String allJars = null;
-
-    private static void buildJar() throws Exception {
-        testJar = JarBuilder.build("test", TEST_CLASSES);
-        bootJar = JarBuilder.build("boot", BOOT_CLASSES);
-        System.out.println("Jar file created: " + testJar);
-        System.out.println("Jar file created: " + bootJar);
-        allJars = testJar+ File.pathSeparator + bootJar;
-    }
-
-    private static boolean argsContain(String[] args, String flag) {
-         for (String s: args) {
-             if (s.contains(flag)) {
-                 return true;
-             }
-         }
-         return false;
-    }
-
-    private static boolean argsContainOpts(String[] args, String... opts) {
-        boolean allIn = true;
-        for (String f : opts) {
-            allIn &= argsContain(args, f);
-            if (!allIn) {
-                break;
-            }
-        }
-        return allIn;
-    }
-
-    private static LingeredApp createLingeredApp(String... args) throws Exception {
-        JCmdTestLingeredApp app  = new JCmdTestLingeredApp();
-        try {
-            LingeredApp.startAppExactJvmOpts(app, args);
-        } catch (Exception e) {
-            // Check flags used.
-            if (argsContainOpts(args, new String[] {"-Xshare:off", "-XX:+RecordDynamicDumpInfo"}) ||
-                argsContainOpts(args, new String[] {"-XX:+RecordDynamicDumpInfo", "-XX:ArchiveClassesAtExit="})) {
-                // app exit premature due to incompactible args
-                return null;
-            }
-            Process proc = app.getProcess();
-            if (e instanceof IOException && proc.exitValue() == 0) {
-                // Process started and exit normally.
-                return null;
-            }
-            throw e;
-        }
-        return app;
-    }
-
-    static int logFileCount = 0;
-    private static void runWithArchiveFile(String archiveName, boolean useBoot,  String... messages) throws Exception {
-        List<String> args = new ArrayList<String>();
-        if (useBoot) {
-            args.add("-Xbootclasspath/a:" + bootJar);
-        }
-        args.add("-cp");
-        if (useBoot) {
-            args.add(testJar);
-        } else {
-            args.add(allJars);
-        }
-        args.add("-Xshare:on");
-        args.add("-XX:SharedArchiveFile=" + archiveName);
-        args.add("-Xlog:class+load");
-
-        LingeredApp app = createLingeredApp(args.toArray(new String[0]));
-        app.setLogFileName("JCmdTestDynamicDump.log." + (logFileCount++));
-        app.stopApp();
-        String output = app.getOutput().getStdout();
-        if (messages != null) {
-            for (String msg : messages) {
-                if (!output.contains(msg)) {
-                    throw new RuntimeException(msg + " missed from output");
-                }
-            }
-        }
-    }
-
-    private static void test(String archiveFile, long pid,
-                             boolean useBoot, boolean expectOK, String... messages) throws Exception {
-        System.out.println("Expected: " + (expectOK ? "SUCCESS" : "FAIL"));
-        String fileName = archiveFile != null ? archiveFile :
-            ("java_pid" + pid + "_dynamic.jsa");
-        File file = new File(fileName);
-        if (file.exists()) {
-            file.delete();
-        }
-
-        String jcmd = "VM.cds dynamic_dump";
-        if (archiveFile  != null) {
-          jcmd +=  " " + archiveFile;
-        }
-
-        PidJcmdExecutor cmdExecutor = new PidJcmdExecutor(String.valueOf(pid));
-        OutputAnalyzer output = cmdExecutor.execute(jcmd, true/*silent*/);
-
-        if (expectOK) {
-            output.shouldHaveExitValue(0);
-            if (!file.exists()) {
-                throw new RuntimeException("Could not create shared archive: " + fileName);
-            } else {
-                runWithArchiveFile(fileName, useBoot, messages);
-                file.delete();
-            }
-        } else {
-            if (file.exists()) {
-                throw new RuntimeException("Should not create shared archive " + fileName);
-            }
-        }
-    }
-
-    private static void print2ln(String arg) {
-        System.out.println("\n" + arg + "\n");
-    }
-
-    private static void test_dynamic() throws Exception {
+    static void test() throws Exception {
+        setIsStatic(false);
         int  test_count = 1;
         final boolean useBoot = true;
         final boolean noBoot = !useBoot;
@@ -195,6 +67,8 @@ public class JCmdTestDynamicDump {
         LingeredApp app  = null;
         long pid;
 
+        // build test.jar and boot.jar
+        buildJars();
         // Test dynamic dump with -XX:+RecordDynamicDumpInfo.
         print2ln(test_count++ + " Test dynamic dump with -XX:+RecordDynamicDumpInfo.");
         app = createLingeredApp("-cp", allJars, "-XX:+RecordDynamicDumpInfo");
@@ -275,16 +149,11 @@ public class JCmdTestDynamicDump {
             .shouldHaveExitValue(0);
         File file = new File(archiveFile);
         if (!file.exists()) {
-            throw new RuntimeException("Shared archive file " + archiveFile + " is not created");
+            throw new RuntimeException("Cannot dump classes to archive file " + archiveFile);
         }
     }
 
     public static void main(String... args) throws Exception {
-        boolean cdsEnabled = WhiteBox.getWhiteBox().getBooleanVMFlag("UseSharedSpaces");
-        if (!cdsEnabled) {
-            throw new SkippedException("CDS is not available for this JDK.");
-        }
-        buildJar();
-        test_dynamic();
+        runTest(JCmdTestDynamicDump::test);
     }
 }
